+++ packages/api-client/src/components/AddressBar/index.ts
@@ -1,3 +1,1 @@
 export { default as AddressBar } from './AddressBar.vue'
-export { default as AddressBarServer } from './AddressBarServer.vue'
-export { default as AddressBarServerItem } from './AddressBarServerItem.vue'
