+++ packages/api-client/src/components/Server/index.ts
@@ -1,1 +1,2 @@
 export { default as ServerVariablesForm } from './ServerVariablesForm.vue'
+export { default as ServerDropdown } from './ServerDropdown.vue'
